requirements.txt
beautifulsoup4
requests
Pillow
mysql-connector-python
google-cloud-storage